# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BookIt SUT API",
    "description": "System Under Test for the BookIt automation framework. All state is namespaced by the X-Test-Session header; X-Test-Now overrides the clock for one request; X-Platform (web, wap, android, ios) enables server-side feature gating.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/bookings": {
      "post": {
        "tags": [
          "bookings"
        ],
        "summary": "Create Booking",
        "operationId": "create_booking_bookings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Booking"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/me/bookings": {
      "get": {
        "tags": [
          "bookings"
        ],
        "summary": "My Bookings",
        "operationId": "my_bookings_me_bookings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Booking"
                  },
                  "type": "array",
                  "title": "Response My Bookings Me Bookings Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/bookings/{booking_id}": {
      "delete": {
        "tags": [
          "bookings"
        ],
        "summary": "Cancel Booking",
        "operationId": "cancel_booking_bookings__booking_id__delete",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/bookings/{booking_id}/ics": {
      "get": {
        "tags": [
          "bookings"
        ],
        "summary": "Export Ics",
        "operationId": "export_ics_bookings__booking_id__ics_get",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One VEVENT (docs/prd/ics-export.md)",
            "content": {
              "text/calendar": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Booking": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "class_id": {
            "type": "string",
            "title": "Class Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "can_cancel": {
            "type": "boolean",
            "title": "Can Cancel"
          },
          "cancel_deadline": {
            "type": "string",
            "format": "date-time",
            "title": "Cancel Deadline"
          },
          "checked_in_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked In At"
          },
          "studio_class": {
            "$ref": "#/components/schemas/StudioClass"
          }
        },
        "type": "object",
        "required": [
          "id",
          "class_id",
          "user_id",
          "created_at",
          "can_cancel",
          "cancel_deadline",
          "checked_in_at",
          "studio_class"
        ],
        "title": "Booking"
      },
      "BookingCreate": {
        "properties": {
          "class_id": {
            "type": "string",
            "title": "Class Id"
          }
        },
        "type": "object",
        "required": [
          "class_id"
        ],
        "title": "BookingCreate"
      },
...
      "ErrorBody": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorBody"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
...
      "StudioClass": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "studio_id": {
            "type": "string",
            "title": "Studio Id"
          },
          "studio_name": {
            "type": "string",
            "title": "Studio Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "instructor": {
            "type": "string",
            "title": "Instructor"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          },
          "duration_min": {
            "type": "integer",
            "title": "Duration Min"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "spots_left": {
            "type": "integer",
            "title": "Spots Left"
          },
          "is_full": {
            "type": "boolean",
            "title": "Is Full"
          },
          "has_started": {
            "type": "boolean",
            "title": "Has Started"
          },
          "is_outdoor": {
            "type": "boolean",
            "title": "Is Outdoor"
          },
          "is_anchor": {
            "type": "boolean",
            "title": "Is Anchor"
          },
          "my_booking_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "My Booking Id",
            "description": "The caller's booking on this class, if any"
          },
          "waitlist_count": {
            "type": "integer",
            "title": "Waitlist Count"
          },
          "my_waitlist_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "My Waitlist Position",
            "description": "The caller's 1-based waitlist position, if any"
          }
        },
        "type": "object",
        "required": [
          "id",
          "studio_id",
          "studio_name",
          "name",
          "category",
          "instructor",
          "start_at",
          "end_at",
          "duration_min",
          "capacity",
          "spots_left",
          "is_full",
          "has_started",
          "is_outdoor",
          "is_anchor",
          "my_booking_id",
          "waitlist_count",
          "my_waitlist_position"
        ],
        "title": "StudioClass"
      },
...
    }
  }
}